Add CheckAddress api

+++ Data/ViewModel/Order/LIST_ITEM_PAYLOAD.cs
@@ -3,7 +3,7 @@
     public class LIST_ITEM_PAYLOAD
     {
         public string PRODUCT_NAME { get; set; }
-        public double PRODUCT_PRICE { get; set; }
+        public int PRODUCT_PRICE { get; set; }
         public int PRODUCT_QUANTITY { get; set; }
     }
 }
